% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asmatrixccf21.R
\name{as.matrix.ccf}
\alias{as.matrix.ccf}
\title{as.matrix.ccf
Convert object of type \code{ccf} into a matrix.}
\usage{
\method{as.matrix}{ccf}(x, rownames.force = TRUE, ...)
}
\arguments{
\item{x}{a \code{ccf} object.}

\item{rownames.force}{logical indicating if the resulting matrix
should have character (rather than \code{NULL}) rownames. The default is
\code{TRUE}.}

\item{...}{additional arguments to be passed to or from methods (unused).}
}
\value{
Numeric matrix
}
\description{
as.matrix.ccf
Convert object of type \code{ccf} into a matrix.
}
\examples{
result <- ccf( 1:10, 1:10, shiftaction = "cut", lag.max = 5,
               plot = FALSE, ci = 0.95)
}
\author{
Jan Seifert
}
