% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccf21.R
\name{ccf}
\alias{ccf}
\title{ccf
\code{ccf} computes the cross-correlation or cross-covariance of two univariate
series.}
\usage{
ccf(
  x,
  y,
  lag.max = NULL,
  type = c("correlation", "covariance"),
  stationary = NULL,
  shiftaction = c("cut", "wrap", "replace", "imprison"),
  replaceby = NULL,
  ci = NULL,
  plot = TRUE,
  na.action = stats::na.fail,
  ...
)
}
\arguments{
\item{x, y}{a univariate numeric vector or time series object.}

\item{lag.max}{maximum lag at which to calculate the acf.
Default is \code{10*log10(N/m)} where \code{N} is the number of observations
and \code{m} the number of series. Will be automatically limited
to one less than the number of observations in the series.}

\item{type}{character string giving the type of \code{ccf} to be
computed. Allowed values are "correlation" (the default) or "covariance".
Will be partially matched.}

\item{stationary}{Shall vectors be treated as if they were a stationary time
series. If \code{FALSE} plain correlations will be used. If \code{x} and \code{y} are both
\code{\link[stats:ts]{time series objects}}, \code{TRUE} is the default, otherwise it is \code{FALSE}.}

\item{shiftaction}{How to handle values that are shifted out of range.
See details.}

\item{replaceby}{A numeric value that is used when \code{shiftaction = "replace"}.
Otherwise it is ignored, but if replacement is set this parameter is mandatory.}

\item{ci}{Confidence level (0 < ci < 1).
Ignored, if \code{type == "covariance"}.}

\item{plot}{logical. If \code{TRUE} (the default) the \code{ccf} is plotted.}

\item{na.action}{function to be called to handle missing values.
\code{na.pass} can be used.}

\item{...}{further arguments to be passed to \code{plot}.}
}
\value{
An S3 object of class "\code{ccf}" and "\code{acf}", which is a list with
the following elements:
\describe{
\item{lag}{A three dimensional array containing the lags at
which the \code{ccf} is estimated. Values for ccf are stored in \code{lag[,,1]}.
This is due to compatibility reasons.}
\item{acf}{An array with the same dimensions as \code{lag} containing
the estimated ccf.}
\item{type}{The type of correlation (same as the \code{type} argument).}
\item{n.used}{The number of observations in sequence.}
\item{series}{The name of the series x. Is always \code{X} and can be ignored.
It has been kept due to compatibility reasons.}
\item{snames}{The series names.}
\item{acf.n}{Vectors with the number of samples used for each
cross-correlation at each lag.}
\item{acf.ci}{Two-column matrix with lower (column 1) and upper
confidence level (column 2) at each lag.}
\item{ci.level}{The set confidence level used to determin \code{acf.ci}.}
\item{shiftaction}{The action that handled values being moved
out of range (same as the \code{shiftaction} argument).}
\item{stationary}{Are the cross-correlations in this structure
based on stationarity assumption. \code{TRUE}/\code{FALSE} (same as the \code{stationary}
argument).}
\item{replacedby}{The value that has been used in case of \code{shiftaction = "replace"} (same as the \code{replaceby} argument).}
}
The lag \code{k} value returned by \code{ccf(x, y)} estimates the correlation
between \code{x[t+k]} and \code{y[t]}.

The result is returned invisibly if plot is TRUE.
}
\description{
ccf
\code{ccf} computes the cross-correlation or cross-covariance of two univariate
series.
}
\details{
Unlike the classic ccf/acf (see \verb{[stats::plot.acf]}) functions
this version does not return a confidence interval around zero but around the
correlation at each lag.
}
\seealso{
\verb{[stats::ccf()]}
}
\author{
Jan Seifert
}
