% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccf21.R
\name{.Cor_ccf}
\alias{.Cor_ccf}
\title{.Cor_ccf
Correlating two vectors x and y. Mean, standard deviation and sample size can
deviate from sample estimates.}
\usage{
.Cor_ccf(
  x,
  y,
  type = c("correlation", "covariance"),
  n = NA,
  mean = c(NA, NA),
  sd = c(NA, NA)
)
}
\arguments{
\item{x, y}{Numeric vectors}

\item{type}{Character string giving the type of acf to be computed.
Allowed values are "correlation" (the default), "covariance". Can be abbreviated.}

\item{n}{Sample size of x and y.}

\item{mean}{Means of x and y. Either a single value for both or one mean each.}

\item{sd}{Standard deviations of x and y. Either a single value for
both or one mean each.}
}
\value{
The correlation between x and y
}
\description{
.Cor_ccf
Correlating two vectors x and y. Mean, standard deviation and sample size can
deviate from sample estimates.
}
\details{
If mean, sd, or n are given, treat x and y as stationary time
series. That implies that you must provide these for the complete sequence
and not for the given \strong{partial} sequences \code{x} and \code{y}.
To be treated as non-stationary, n, mean and sd must ALL be NA.
In this case, the functions determines mean, sd, and n based on x and y.
}
\seealso{
\code{\link[stats:cor]{stats::cor()}} \code{\link[stats:cor]{stats::cov()}}
}
\author{
Jan Seifert
}
