% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccf21.R
\name{.CorConf_Fisher}
\alias{.CorConf_Fisher}
\title{.CorConf_Fisher
Computes confidence intervals for correlation coefficients.}
\usage{
.CorConf_Fisher(r, n, ci = 0.95)
}
\arguments{
\item{r}{Correlation coefficient (pearson)}

\item{n}{Sample size of the data underlying r}

\item{ci}{Confidence limit. Default is 0.95.}
}
\value{
A two-column matrix with each row holding the range of confidence.
}
\description{
.CorConf_Fisher
Computes confidence intervals for correlation coefficients.
}
\details{
The function uses the Fisher z transform to get approximately normal
distributed correlations, computes the confidence range and transforms the
values back to correlations.
}
\note{
Note: these confidence intervals are not symmetrical because of
the characteristics of the probability distribution of correlations.
}
\references{
Bonett, D. G. & Wright, T. A. (2000). Sample Size Requirements
for Estimating Pearson, Kendall and Spearman Correlations.” Psychometrika,
65 (1), p. 23-28.
}
\author{
Jan Seifert
}
