% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccfmax.R
\name{max.ccf}
\alias{max.ccf}
\title{max.ccf}
\usage{
\method{max}{ccf}(x, tol = 0)
}
\arguments{
\item{x}{An object of class \code{acf} as returned from \code{stats::ccf()} or \code{ccf} as
returned from the ccf21 implementation.}

\item{tol}{A tolerance below the maximum. When \code{tol != 0} not only the absolute
maximum will be returned but all values within the interval  $\link{max, max-tol}$.}
}
\value{
A matrix with two columns. First column contains the correlation / covariances.
The second column gives the according lags.
}
\description{
max.ccf
}
