% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotccf21.R
\name{plot.ccf}
\alias{plot.ccf}
\title{plot.ccf
Plot method for objects of class "\code{ccf}".}
\source{
\url{https://stackoverflow.com/questions/28857241/r-combine-plots-that-use-parmfrow-internally}
by Alex A. (\url{https://stackoverflow.com/users/3005513/alex-a})
}
\usage{
\method{plot}{ccf}(
  x,
  ci = 0.95,
  type = "h",
  xlab = "Lag",
  ylab = NULL,
  sub = NULL,
  ylim = NULL,
  main = NULL,
  ci.col = "#3366FF",
  ci.type = c("white", "ma"),
  cex.main = 1,
  mar = NULL,
  max.mfrow = NULL,
  ask = NULL,
  oma = NULL,
  mgp = NULL,
  xpd = NULL,
  verbose = NULL,
  ...
)
}
\arguments{
\item{x}{an object of class \code{ccf}.}

\item{ci}{coverage probability for confidence interval.
Plotting of the confidence interval is suppressed if \code{ci} is zero or negative.}

\item{type}{the type of plot to be drawn, default to histogram like
vertical lines. See \verb{[plot()]} for details.}

\item{xlab}{the x label of the plot.}

\item{ylab}{the y label of the plot.}

\item{sub}{subtitle for the plot.}

\item{ylim}{numeric of length 2 giving the y limits for the plot.}

\item{main}{overall title for the plot.}

\item{ci.col}{colour to plot the confidence interval lines.}

\item{ci.type}{should the confidence limits assume a white noise
input or for lag \code{k} an \code{MA(k-1)} input? Can be abbreviated.}

\item{cex.main}{Default size for the title will be multiplied with \code{cex.main}
to adjust the size.}

\item{mar}{Numerical vector of the form \code{c(bottom, left, top, right)}
which gives the number of text lines of margin.}

\item{max.mfrow}{Unused. Available only for compatibility reasons.}

\item{ask}{Unused. Available only for compatibility reasons.}

\item{oma}{Unused. Available only for compatibility reasons.}

\item{mgp}{Unused. Available only for compatibility reasons.}

\item{xpd}{Unused. Available only for compatibility reasons.}

\item{verbose}{Unused. Available only for compatibility reasons.}

\item{...}{Unused. Available only for compatibility reasons.}
}
\description{
plot.ccf
Plot method for objects of class "\code{ccf}".
}
\note{
The white noise option for confidence intervals is probably misleading.
It is based on an uncorrelated series and should be treated with appropriate caution.
It is only kept here for compatibility reasons.

This implementation drops several arguments. They stay only for
only for compatibility reasons but will be ignored.
}
\seealso{
\verb{[stats::plot.acf()]} \verb{[ggplot2::ggplot()]}
\verb{[grapics::plot()]} for arguments: \code{type}, \code{mar}.
}
\author{
Jan Seifert
}
