% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printccf21.R
\name{print.ccf}
\alias{print.ccf}
\title{print.ccf
Prints a cross-correlation object.}
\usage{
\method{print}{ccf}(x, digits = max(3, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an object to be printed.}

\item{digits}{minimal number of significant digits}

\item{...}{further arguments passed to print methods.}
}
\description{
print.ccf
Prints a cross-correlation object.
}
\examples{
print(ccf(1:10, 10:1))
}
\seealso{
\code{\link[base:print.dataframe]{base::print.data.frame()}}
}
\author{
Jan Seifert
}
