% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryccf21.R
\name{print.summaryccf}
\alias{print.summaryccf}
\title{print.summaryccf
Prints the summary of a \code{ccf} object.}
\usage{
\method{print}{summaryccf}(x, digits = max(3, getOption("digits") - 3L))
}
\arguments{
\item{x}{an object of class "\code{summaryccf}", usually, a result of
a call to \code{summary.ccf}.}

\item{digits}{the number of significant digits to use when printing.}
}
\description{
print.summaryccf
Prints the summary of a \code{ccf} object.
}
\examples{
print( summary(ccf(1:10, 10:1)) )
}
\seealso{
\verb{[base::print()]}
}
\author{
Jan Seifert
}
