% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryccf21.R
\name{summary.ccf}
\alias{summary.ccf}
\title{summary.ccf
Summary for a cross-correlation object \code{ccf}.}
\usage{
\method{summary}{ccf}(object, ...)
}
\arguments{
\item{object}{an object of class \code{ccf}.}

\item{...}{additional arguments to be passed to or from methods (unused).}
}
\value{
This function computes and returns a list of summary statistics:
\describe{
\item{type}{Used statistics correlation or covariance.}
\item{smethod}{The shifting method used in call to \code{ccf}.}
\item{lagrange}{The range of the lags.}
\item{nacount}{The number of NAs in the statistics.}
\item{range}{The range of values in sequence.}
\item{min}{Minimum correlation/covariance found.}
\item{minpos}{Index positions where the maxima were found.}
\item{minci}{Lower confidence limit at \code{minpos}.}
\item{max}{Maximum correlation/covariance found.}
\item{maxpos}{Index positions where the maxima were found.}
\item{maxci}{Upper confidence limit at \code{maxpos}.}
}
}
\description{
summary.ccf
Summary for a cross-correlation object \code{ccf}.
}
\examples{
summary(ccf(1:10, 10:1))
}
\seealso{
\verb{[base::summary()]}
}
\author{
Jan Seifert
}
